Limit sysv_ipc to non-Windows installs
This dependency is compatible with "most (all?) Unixes", Windows support
is unofficial and only under Cygwin. Cygwin and WSL Python installs have
values of `platform.system()` other than Windows, therefore this prevents
installlation of `sysv_ipc` under standard Windows. This will cause `PulseIn`
to generate import errors on Windows, but as there is no Windows driver this
is likely not a problem. It will, however, allow for dependency resolution
and packaging to work on Windows.