Require Jetson.GPIO on aarch64
The Tegra pin.py files depend on the Jetson.GPIO library. Currently, they
assume the library is installed in /opt/nvidia/jetson-gpio and hard-code
this path. This doesn't work well when someone has installed an upgraded
version, e.g. via PYTHONPATH or pip. To solve this, remove the
hard-coding of Jetson.GPIO's installation path, and update setup.py and
requirements.txt to ensure that Jetson.GPIO is available. Note that
Jetson.GPIO is now available via PyPi, but may not have been when initial
Jetson support was added to Blinka.