Add PWMOut for PocketBeagle #159
Add PWMOut for PocketBeagle which has the AM3358 SoC.

Note that the generic linux PWMOut implementation will
not work as BeagleBoard.org kernel has a patch to allow
udev to be able to set the ownership and permissions
in /sys/class/pwm when channels are exported.  However,
this naming scheme is different than what is used in the
mainline kernel.

The PocketBeagle has these PWM outputs:

/sys/class/pwm/pwmchip0/pwm-0:0/period
/sys/class/pwm/pwmchip0/pwm-0:1/period
/sys/class/pwm/pwmchip2/pwm-2:0/period
/sys/class/pwm/pwmchip2/pwm-2:1/period
/sys/class/pwm/pwmchip4/pwm-4:0/period
/sys/class/pwm/pwmchip4/pwm-4:1/period

Here is the kernel patch that is being used:
https://github.com/RobertCNelson/bb-kernel/blob/am33x-v4.11/patches/drivers/pwm/0001-pwm-Create-device-class-for-pwm-channels.p$