Add support for ftdi URLs in the BLINKA_FT232H variable.
This allows the an FTDI URL, like ftdi://ftdi:232h:W00001/1
to be given in the BLINKA_FT232H. If it's used, then that
board is the one that is opened. If the value of the variable
is not an FTDI url, then the default is used.

This makes it possible for multiple processes on the same
computer to talk to different FTDI boards. While talking
to multiple boards from a single process might be possible,
that's outside of the scope of this commit.